pyproject.toml
[project]
name = "the-youtuber"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-functions>=1.24.0",
    "fastapi>=0.123.9",
    "google-genai>=1.53.0",
    "google-generativeai>=0.8.5",
    "ipykernel>=7.1.0",
    "lancedb>=0.25.3",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.27.0",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.52.0",
    "uvicorn>=0.38.0",
]
